requirements.txt
langchain
langchain-qdrant
sentence_transformers
transformers
torch
pypdf
qdrant-client
google-generativeai
python-dotenv
fastapi
uvicorn
groq
openai
pandas